# docs/api_spec_full.yaml
openapi: 3.0.3
info:
  title: SoulRep Gym Management API
  description: |
    Complete API suite for SoulRep Gym.
    Covers Auth, Member Dashboard, Owner Analytics, and Trainer Operations.
  version: 2.1.0
servers:
  - url: http://localhost:3000/api
paths:
  /auth/reset-password:
    post:
      summary: Reset user password
      description: Verifies user by email and phone to reset password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, phone, newPassword]
              properties:
                email: { type: string }
                phone: { type: string }
                newPassword: { type: string }
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '400': { description: Missing required fields }
        '404': { description: No account found }
...
  /member/bookings:
    get:
      summary: Get member bookings
      parameters:
        - name: upcoming
          in: query
          schema: { type: boolean }
        - name: status
          in: query
          schema: { type: string, enum: [PENDING, CONFIRMED, COMPLETED, CANCELLED] }
        - name: trainerId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Booking' } }
    post:
      summary: Create session booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trainerId, date, time]
              properties:
                trainerId: { type: string }
                date: { type: string, format: date }
                time: { type: string, example: "09:00" }
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '403': { description: No sessions remaining }
...
  /member/nutrition:
    get:
      summary: Get nutrition logs and AI generated diet plan
      responses:
        '200':
          description: Current nutrition profile and active diet plans
          content:
            application/json:
              schema: { type: object }
    post:
      summary: Update nutrition parameters and trigger AI generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                age: { type: integer, nullable: true }
                weight: { type: number, nullable: true }
                height: { type: number, nullable: true }
                fitnessGoal: { type: string }
                activityLevel: { type: string }
                dietaryPreference: { type: string }
                cuisinePreference: { type: string }
                usualDiet: { type: string }
                allergies: { type: array, items: { type: string } }
                restrictions: { type: array, items: { type: string } }
      responses:
        '200':
          description: AI generation complete or in progress
          content:
            application/json:
              schema: { type: object }
...
  /owner/equipment:
    get:
      summary: List workout equipment
      responses:
        '200':
          description: Equipment list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Equipment' }
    post:
      summary: Add new equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, category]
              properties:
                name: { type: string }
                category: { type: string }
                status: { type: string }
                purchaseDate: { type: string, format: date }
      responses:
        '201': { description: Created }
    patch:
      summary: Update equipment status/maintenance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string }
                status: { type: string }
                maintenance:
                  type: object
                  properties:
                    type: { type: string }
                    description: { type: string }
                    cost: { type: number }
                    performedBy: { type: string }
      responses:
        '200': { description: Updated }
...
  /owner/invoices:
    get:
      summary: List all invoices
      responses:
        '200': { description: Invoices }
    post:
      summary: Create manual invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [memberId, plan, amount]
              properties:
                memberId: { type: string }
                plan: { type: string }
                amount: { type: number }
      responses:
        '201': { description: Created }
...
  /owner/members:
    get:
      summary: List all gym members
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Member' }
    patch:
      summary: Update member details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [memberId]
              properties:
                memberId: { type: string }
                plan: { type: string }
                trainerId: { type: string }
                planStatus: { type: string }
                sessionsRemaining: { type: integer }
                planExpiry: { type: string, format: date }
      responses:
        '200': { description: Updated }
    delete:
      summary: Delete member account
      parameters:
        - name: memberId
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: Deleted }
...
  /owner/trainers/{trainerId}/members:
    get:
      summary: List members assigned to trainer
      parameters:
        - name: trainerId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: List of members }
  /owner/trainers/{trainerId}/reviews:
    get:
      summary: Get reviews for specific trainer
      parameters:
        - name: trainerId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Reviews }
...
components:
  schemas:
...
    Member:
      type: object
      properties:
        id: { type: string }
        joinDate: { type: string, format: date-time }
        plan: { type: string, enum: [BASIC, PRO, ELITE] }
        planStatus: { type: string, enum: [ACTIVE, EXPIRED, CANCELLED] }
        sessionsRemaining: { type: integer }
        age: { type: integer }
        gender: { type: string }
        trainerId: { type: string }
...
    Booking:
      type: object
      properties:
        id: { type: string }
        trainerId: { type: string }
        trainerName: { type: string }
        date: { type: string, format: date-time }
        time: { type: string }
        status: { type: string, enum: [PENDING, CONFIRMED, COMPLETED, CANCELLED] }
...
    Equipment:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        status: { type: string }
        lastMaintenance: { type: string, format: date-time }